APPS-3146 fix parsing exception of color

+++ common/src/main/java/org/geogebra/common/awt/GColor.java
@@ -249,10 +249,10 @@ public final class GColor implements GPaint {
 	 * @return gColor
 	 */
 	public static GColor newColor(String r, String g, String b, String a) {
-		int red = Integer.parseInt(r);
-		int green = Integer.parseInt(g);
-		int blue = Integer.parseInt(b);
-		int alpha = Integer.parseInt(a);
+		double red = Double.parseDouble(r);
+		double green = Double.parseDouble(g);
+		double blue = Double.parseDouble(b);
+		double alpha = Double.parseDouble(a);
 		return newColor(red, green, blue, alpha * 255);
 	}
